Add check for the http status code returned by the tracker
The `bencode.decode()` would sometimes throw an error in a callback without propagating that error to the callback of `tracker.scrape()` effectively resulting in the termination of the node.js process unless user had `process.on("uncaughtException", ...)` listener set in their application.

In this commit I added an early check for the unsuccessful http response before attempting to parse its body. I also added a try-catch for `bencode.decode()` with proper propagation of the error just in case tracker returns garbage body despite the successful response code.